unify vue app naming clean dependencies for every app clean scripts cleanup packages.json, names, version, descriptions... add dev server for ocis and oc10 move oc10 php files and appinfo to own package
