+++ .depcheckrc
@@ -1,0 +1,9 @@
+{
+  "ignores": [
+    "cucumber-pretty",
+    "requirejs",
+    "nightwatch-vrt",
+    "@babel/polyfill",
+    "@babel/register"
+  ]
+}
