+++ package.json
@@ -10,30 +10,17 @@
   "scripts": {
     "build": "rollup -c",
     "build:w": "rollup -c -w",
-    "dep:test:app": "if [ -d tests/testing-app ]; then (cd tests/testing-app && git pull && pwd); else git clone --depth 1 https://github.com/owncloud/testing.git tests/testing-app; fi;",
+    "depcheck": "depcheck",
     "docker:kill": "docker kill web-tests-redis web-tests-selenium",
     "docker:redis": "docker run --rm -d -p 6379:6379 -e REDIS_DATABASES=1 --name web-tests-redis webhippie/redis:latest",
     "docker:selenium": "docker run --rm -d --network=\"host\" -v /dev/shm:/dev/shm -v ${REMOTE_UPLOAD_DIR:-$PWD/tests/acceptance/filesForUpload}:${LOCAL_UPLOAD_DIR:-/uploads}:ro --name web-tests-selenium selenium/standalone-chrome-debug",
     "docker:selenium:mac": "docker run --rm -d -p ${SELENIUM_PORT:-4444}:4444 -p 5900:5900 -v /dev/shm:/dev/shm -v ${REMOTE_UPLOAD_DIR:-$PWD/tests/acceptance/filesForUpload}:${LOCAL_UPLOAD_DIR:-/uploads}:ro --name web-tests-selenium selenium/standalone-chrome-debug",
     "lint": "eslint packages/*/src tests --ext vue --ext js --color",
-    "lint:fix": "eslint packages/*/src tests --ext vue --ext js --color --fix",
+    "server": "yarn build:w",
     "test:acceptance:drone": "cucumber-js --retry 1 --require-module @babel/register --require-module @babel/polyfill --require tests/acceptance/setup.js --require tests/acceptance/stepDefinitions --format node_modules/cucumber-pretty ${TEST_PATHS:-tests/acceptance/features/$TEST_CONTEXT} -t \"${TEST_TAGS}\"",
     "test:acceptance:oc10": "cucumber-js --require-module @babel/register --require-module @babel/polyfill --require tests/acceptance/setup.js --require tests/acceptance/stepDefinitions --format node_modules/cucumber-pretty -t \"${TEST_TAGS:-not @skip and not @skipOnOC10}\"",
     "test:acceptance:ocis": "NODE_TLS_REJECT_UNAUTHORIZED=0 RUN_ON_OCIS=true cucumber-js --require-module @babel/register --require-module @babel/polyfill --require tests/acceptance/setup.js --require tests/acceptance/stepDefinitions --format node_modules/cucumber-pretty -t \"${TEST_TAGS:-not @skip and not @skipOnOCIS and not @notToImplementOnOCIS}\"",
-    "test:acceptance:prepare": "yarn docker:redis; yarn docker:selenium; yarn dep:test:app",
-    "test:acceptance:prepare:mac": "yarn docker:redis; yarn docker:selenium:mac; yarn dep:test:app",
-    "test:accessibility": "node node_modules/axe-cli/axe-cli --exit --disable page-has-heading-one,meta-viewport"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "*.{vue,js}": [
-      "yarn lint --fix",
-      "git add"
-    ]
+    "test:unit": "jest --coverage --config ./tests/unit/config/jest.config.js"
   },
   "author": "ownCloud",
   "license": "AGPL-3.0",
@@ -46,18 +33,10 @@
     "not IE_Mob 11"
   ],
   "devDependencies": {
-    "@babel/cli": "^7.6.0",
     "@babel/core": "^7.12.10",
-    "@babel/plugin-proposal-class-properties": "^7.5.5",
-    "@babel/plugin-proposal-export-default-from": "^7.5.2",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
-    "@babel/plugin-proposal-optional-chaining": "^7.8.3",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-transform-classes": "^7.5.5",
-    "@babel/plugin-transform-runtime": "^7.6.0",
-    "@babel/polyfill": "^7.8.7",
+    "@babel/polyfill": "^7.12.1",
     "@babel/preset-env": "^7.12.11",
-    "@babel/register": "^7.6.0",
+    "@babel/register": "^7.12.10",
     "@erquhart/rollup-plugin-node-builtins": "^2.1.5",
     "@rollup/plugin-alias": "^3.1.1",
     "@rollup/plugin-commonjs": "^17.0.0",
@@ -65,16 +44,12 @@
     "@rollup/plugin-json": "^4.1.0",
     "@vue/test-utils": "^1.1.2",
     "archiver": "^3.0.0",
-    "axe-cli": "^3.1.0",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.2",
     "babel-jest": "^26.6.3",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-transform-es2015-arrow-functions": "^6.22.0",
-    "babel-plugin-transform-es2015-shorthand-properties": "^6.24.1",
     "chromedriver": "^87.0.0",
     "cucumber": ">=6.0.5",
     "cucumber-pretty": ">=6.0.0",
+    "depcheck": "^1.3.1",
     "ejs": "^3.1.5",
     "eslint": "6.8.0",
     "eslint-config-prettier": "^6.10.0",
@@ -88,25 +63,16 @@
     "eslint-plugin-vue": "^6.2.2",
     "fs-extra": "^9.0.1",
     "html-minifier": "^4.0.0",
-    "husky": ">=3.0.5",
-    "inert-polyfill": "^0.2.5",
     "jest": "^26.6.3",
     "jest-serializer-vue": "^2.0.2",
     "jest-svg-transformer": "^1.0.0",
     "join-path": "^1.1.1",
     "ldapjs": "^2.2.3",
-    "lint-staged": ">=9",
     "lodash": "^4.17.15",
     "nightwatch": "1.3.4",
     "nightwatch-api": "3.0.1",
     "nightwatch-vrt": "^0.2.10",
     "node-fetch": "^2.6.1",
-    "node-fs-extra": "^0.8.2",
-    "node-sass": "^5.0.0",
-    "npm-run-all": "^4.1.5",
-    "oidc-client": "^1.9.1",
-    "owncloud-design-system": "^2.1.2",
-    "owncloud-sdk": "^1.0.0-740",
     "p-limit": "^2.2.1",
     "postcss": "^8.2.2",
     "requirejs": "^2.3.6",
@@ -120,10 +86,12 @@
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-postcss": "^4.0.0",
     "rollup-plugin-progress": "^1.1.2",
+    "rollup-plugin-serve": "^1.1.0",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-visualizer": "^4.2.0",
     "rollup-plugin-vue": "^5.1.4",
     "url-search-params-polyfill": "^8.0.0",
+    "vue-jest": "^3.0.7",
     "vue-template-compiler": "^2.6.12",
     "xml-js": "^1.6.11"
   },
