--- packages/web-app-draw-io/.gitignore
@@ -1,3 +1,0 @@
-/js/
-/dist/
-/node_modules/
