+++ packages/web-app-draw-io/src/App.vue
