+++ packages/web-app-files/package.json
@@ -1,16 +1,13 @@
 {
   "name": "files",
-  "version": "0.1.0",
-  "description": "This is the funky frontend app",
-  "homepage": "https://github.com/owncloud/web",
+  "version": "1.0.0",
+  "description": "ownCloud web files",
   "license": "AGPL-3.0",
-  "author": "Felix Heidecke",
-  "main": "js/files.bundle.js",
   "devDependencies": {
     "copy-to-clipboard": "^3.3.1",
     "filesize": "^6.1.0",
     "filter-obj": "^2.0.1",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.20",
     "lodash-es": "^4.17.20",
     "moment": "^2.24.0",
     "p-queue": "^6.6.2",
