+++ packages/web-app-media-viewer/src/App.vue
