--- packages/web-container/.htaccess
@@ -1,16 +1,0 @@
-
-# BEGIN Caching
-<filesMatch "\.(html|js|css|json|svg|jpeg)$">
-  FileETag None
-  <ifModule mod_headers.c>
-     Header unset ETag
-     Header set Cache-Control "max-age=0, no-cache, no-store, must-revalidate"
-     Header set Pragma "no-cache"
-     Header set Expires "Wed, 11 Jan 1984 05:00:00 GMT"
-  </ifModule>
-</filesMatch>
-# END Caching
-Header      unset X-Frame-Options
-Header always set X-Frame-Options "DENY"
-Header      unset Content-Security-Policy
-Header always set Content-Security-Policy "frame-ancestors 'none';"
