--- packages/web-container/themes/example.json
@@ -1,6 +1,0 @@
-{
-  "general": {
-    "name": "Custom Cloud",
-    "slogan": "Your custom cloud, personalized for you!"
-  }
-}
