+++ packages/web-integration-oc10/appinfo/info.xml
