+++ packages/web-runtime/package.json
@@ -1,11 +1,8 @@
 {
-  "name": "owncloud_web",
+  "name": "web-runtime",
   "version": "1.0.0",
-  "description": "A safe home for all your data",
-  "homepage": "https://github.com/owncloud/web",
+  "description": "ownCloud web runtime",
   "license": "AGPL-3.0",
-  "author": "ownCloud",
-  "main": "core/core.bundle.js",
   "devDependencies": {
     "inert-polyfill": "^0.2.5",
     "lodash-es": "^4.17.20",
@@ -28,7 +25,5 @@
     "vuex-persist": "2.0.1",
     "vuex-router-sync": "^5.0.0",
     "wicked-good-xpath": "^1.3.0"
-  },
-  "icon": "folder",
-  "title": "ownCloud Web"
+  }
 }
